# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "idp API",
        "contact": {
            "name": "idp Team",
            "email": "idp@overal-x.org"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/network": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "operationId": "listNetworks",
                "parameters": [
                    {
                        "type": "string",
                        "name": "machineId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.NetworkModel"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "operationId": "createNetwork",
                "parameters": [
                    {
                        "description": "List Networks",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateNetworkArgs"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.NetworkModel"
                        }
                    }
                }
            }
        },
        "/network/{networkId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "operationId": "deleteNetwork",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network Id",
                        "name": "networkId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/repo-connection": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo-connection"
                ],
                "operationId": "listRepoConnections",
                "parameters": [
                    {
                        "type": "string",
                        "name": "machineId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "take",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.RepoConnectionModel"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo-connection"
                ],
                "operationId": "createRepoConnection",
                "parameters": [
                    {
                        "description": "Create Repo Connection",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateRepoConnectionArgs"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.RepoConnectionModel"
                        }
                    }
                }
            }
        },
        "/repo-connection/{connectionId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo-connection"
                ],
                "operationId": "getRepoConnection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repo Connection Id",
                        "name": "connectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/db.RepoConnectionModel"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo-connection"
                ],
                "operationId": "deleteRepoConnection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repo Connection Id",
                        "name": "connectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo-connection"
                ],
                "operationId": "updateRepoConnection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repo Connection Id",
                        "name": "connectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Repo Connection",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateRepoConnectionArgs"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/db.RepoConnectionModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Currency": {
            "type": "string",
            "enum": [
                "NGN"
            ],
            "x-enum-varnames": [
                "CurrencyNgn"
            ]
        },
        "db.DeploymentLogModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deployment": {
                    "$ref": "#/definitions/db.DeploymentModel"
                },
                "deploymentId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.DeploymentModel": {
            "type": "object",
            "properties": {
                "DeploymentLog": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentLogModel"
                    }
                },
                "actor": {
                    "type": "string"
                },
                "commitHash": {
                    "type": "string"
                },
                "commitMessage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "repoConnection": {
                    "$ref": "#/definitions/db.RepoConnectionModel"
                },
                "repoConnectionId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/db.DeploymentStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.DeploymentStatus": {
            "type": "string",
            "enum": [
                "IDLE",
                "IN_PROGRESS",
                "SUCCESSFUL",
                "FAILED"
            ],
            "x-enum-varnames": [
                "DeploymentStatusIdle",
                "DeploymentStatusInProgress",
                "DeploymentStatusSuccessful",
                "DeploymentStatusFailed"
            ]
        },
        "db.GithubAccountConnectionModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "githubEmail": {
                    "type": "string"
                },
                "githubId": {
                    "type": "string"
                },
                "githubInstallationId": {
                    "type": "integer"
                },
                "githubUsername": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.UserModel"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "db.MachineModel": {
            "type": "object",
            "properties": {
                "Deployment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentModel"
                    }
                },
                "Network": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.NetworkModel"
                    }
                },
                "RepoConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.RepoConnectionModel"
                    }
                },
                "containerId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machineImage": {
                    "type": "string"
                },
                "machineName": {
                    "type": "string"
                },
                "machineStatus": {
                    "$ref": "#/definitions/db.MachineStatus"
                },
                "owner": {
                    "$ref": "#/definitions/db.UserModel"
                },
                "ownerId": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/db.MachinePlanModel"
                },
                "planId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.MachinePlanModel": {
            "type": "object",
            "properties": {
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "cpu": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/db.Currency"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hourlyRate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "monthlyRate": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.MachineStatus": {
            "type": "string",
            "enum": [
                "CREATING",
                "RUNNING",
                "REBOOTING",
                "SHUTTING_DOWN",
                "SHUTDOWN"
            ],
            "x-enum-varnames": [
                "MachineStatusCreating",
                "MachineStatusRunning",
                "MachineStatusRebooting",
                "MachineStatusShuttingDown",
                "MachineStatusShutdown"
            ]
        },
        "db.NetworkModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "destinationPort": {
                    "type": "integer"
                },
                "hostName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingressId": {
                    "type": "string"
                },
                "listeningPort": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.RepoConnectionModel": {
            "type": "object",
            "properties": {
                "Deployment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "repoId": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.UserModel": {
            "type": "object",
            "properties": {
                "GithubAccountConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GithubAccountConnectionModel"
                    }
                },
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "types.CreateNetworkArgs": {
            "type": "object",
            "properties": {
                "destinationPort": {
                    "type": "integer"
                },
                "listeningPort": {
                    "type": "integer"
                },
                "machineId": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                }
            }
        },
        "types.CreateRepoConnectionArgs": {
            "type": "object",
            "properties": {
                "machineId": {
                    "type": "string"
                },
                "repoId": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateRepoConnectionArgs": {
            "type": "object",
            "properties": {
                "machineId": {
                    "type": "string"
                },
                "repoId": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                }
            }
        },
...
    }
}